backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==5.1.2
django-cors-headers==4.6.0
django-ratelimit==4.1.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
EbookLib==0.18
boto3==1.34.162
gunicorn==23.0.0
packaging==24.2
pillow==11.0.0
psycopg==3.3.2
psycopg2==2.9.11
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.1
tzdata==2024.2
